# Adds gojs.MustInstantiate to avoid conflicts