This separate host from guest instantiation in ways similar to other imports such as emscripten. Doing so allows parallel use of gojs.Run, provided the ModuleConfig has been assigned a unique name (e.g. via an atomic number).

Fixes #939