@codefromthecrypt 2022-12-19
rebased, and benchmarks look as expected where there is less overhead when the "go" module isn't instantiated every run.

```bash
# before change
$ go test -run='^$' -bench '^Benchmark_main$' -count=5 > old.txt
# after change
$ go test -run='^$' -bench '^Benchmark_main$' -count=5 > new.txt
$ benchstat old.txt new.txt
name               old time/op    new time/op    delta
_main/gojs.Run-16    9.65ms ± 4%    9.45ms ± 2%    ~     (p=0.548 n=5+5)

name               old alloc/op   new alloc/op   delta
_main/gojs.Run-16    23.1MB ± 0%    22.9MB ± 0%  -0.72%  (p=0.008 n=5+5)

name               old allocs/op  new allocs/op  delta
_main/gojs.Run-16     80.4k ± 0%     77.7k ± 0%    ~     (p=0.079 n=4+5)
```